requirements.txt
django>=5.1.5
google-generativeai>=0.8.6
python-dotenv>=1.0.1
